requirements.txt
requests
python-dotenv
rich
pdf2image
pdfkit
Pillow
fastapi
uvicorn
jinja2
markdown
bcrypt
python-multipart
pymongo